requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
blis==0.7.11
catalogue==2.0.10
cbor==1.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
coloredlogs==15.0.1
confection==0.1.4
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.9
einops==0.7.0
faiss-cpu==1.8.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
ijson==3.2.3
importlib_resources==6.4.0
inscriptis==2.5.0
ir-datasets==0.5.6
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
lightgbm==4.3.0
lightning==2.2.1
lightning-utilities==0.11.2
lxml==5.1.0
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nmslib==2.1.1
numpy==1.25.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.17.1
packaging==24.0
pandas==2.0.3
pillow==10.2.0
preshed==3.0.9
protobuf==5.26.0
psutil==5.9.8
pyautocorpus==0.1.12
pybind11==2.6.1
pydantic==2.6.4
pydantic_core==2.16.3
pyjnius==1.6.1
pyparsing==3.1.2
pyserini==0.20.0
python-dateutil==2.8.2
pytorch-lightning==2.2.1
pytrec-eval==0.5
pytz==2024.1
PyYAML==6.0.1
qwikidata==0.4.2
rapidfuzz==3.7.0
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.9.3
seaborn==0.13.2
sentence-transformers==2.6.1
sentencepiece==0.2.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thefuzz==0.22.1
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.1
torchmetrics==1.3.2
torchvision==0.17.1
tqdm==4.66.2
transformers==4.39.0
trec-car-tools==2.6
triton==2.2.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
unlzw3==0.2.2
urllib3==2.2.1
warc3-wet==0.2.3
warc3-wet-clueweb09==0.2.5
wasabi==1.1.2
weasel==0.3.4
yarl==1.9.4
zipp==3.18.1
zlib-state==0.1.6
